refactor(mcp): move mcp transforms into internal/server/transform package
- Move transform_mcp_tool_injection, transform_mcp_tool_strip_guard
  and their tests into internal/server/transform.
- Extract mergeUnique helpers from mcp_integration.go into the
  new package as mcp_merge_tools.go.
- Decouple transforms from *server.Server by accepting *runtime.Runtime
  directly to avoid circular imports.
- Update chain_builder.go to wire the new package.
