+++ internal/server/chain_builder.go
@@ -30,9 +30,6 @@ func (s *Server) BuildTransformChain(c *gin.Context, targetType protocol.APIType
 	// 2. Base transform (protocol conversion)
 	transforms = append(transforms, transform.NewBaseTransform(targetType))
 	if s.mcpEnabled() {
-	}
-	transforms = append(transforms, transform.NewVendorTransform(providerURL))
-
 	// 3. Post-transform recording (if request recording is enabled)
 	if shouldRecord && (recordMode == obs.RecordModeAll || recordMode == obs.RecordModeScenario) {
 		transforms = append(transforms, NewPostTransformRecorder(recorder, c))
